moved dashboard to home
dashboard has been moved to homepage, and redirects to login if not logged in yet.  I also noticed that manage profile was not working, so I made a fix in the controller to make it work.
